chore: optimize UI
