+++ app/src/main/java/io/github/usharerose/flexpal/gamepad/android/MainActivity.kt
@@ -21,7 +21,7 @@ class MainActivity : AppCompatActivity() {
 
         slidersContainer = findViewById<LinearLayout>(R.id.slidersContainer)
         for (i in 1..CHAMBER_COUNT) {
-            appendChamberInputView()
+            appendChamberPressureInputView(i)
         }
 
         valuesTextView = findViewById<TextView>(R.id.valuesTextView)
@@ -33,8 +33,8 @@ class MainActivity : AppCompatActivity() {
         udpManager.release()
     }
 
-    private fun appendChamberInputView() {
-        val sliderEditTextView = ChamberInputView(this).apply {
+    private fun appendChamberPressureInputView(idx: Int) {
+        val chamberPressureInputView = ChamberPressureInputView(this).apply {
             layoutParams = LinearLayout.LayoutParams(
                 LinearLayout.LayoutParams.MATCH_PARENT,
                 LinearLayout.LayoutParams.WRAP_CONTENT
@@ -42,14 +42,15 @@ class MainActivity : AppCompatActivity() {
                 setMargins(0, 8, 0, 8)
             }
 
-            setOnValueChangeListener(object : ChamberInputView.OnValueChangeListener {
+            setOnValueChangeListener(object : ChamberPressureInputView.OnValueChangeListener {
                 override fun onValueChanged() {
                     updateValuesDisplay()
                 }
             })
         }
+        chamberPressureInputView.setChamberIndex(idx)
 
-        slidersContainer.addView(sliderEditTextView)
+        slidersContainer.addView(chamberPressureInputView)
     }
 
     private fun updateValuesDisplay() {
@@ -65,7 +66,7 @@ class MainActivity : AppCompatActivity() {
 
     private fun getAllValues(): List<Int> {
         return (0 until slidersContainer.childCount).map { index ->
-            val chamberInputView = slidersContainer.getChildAt(index) as ChamberInputView
+            val chamberInputView = slidersContainer.getChildAt(index) as ChamberPressureInputView
             chamberInputView.getValue()
         }
     }
