+++ app/src/main/res/layout/activity_main.xml
@@ -11,6 +11,8 @@
         android:textSize="24sp"
         android:textStyle="bold"
         android:layout_marginBottom="16dp"
+        android:layout_marginTop="24dp"
+        android:layout_marginStart="16dp"
         android:gravity="center" />
 
     <!-- <include
@@ -27,5 +29,8 @@
         android:layout_height="wrap_content"
         android:text="Values: "
         android:textSize="18sp"
-        android:layout_marginBottom="16dp" />
+        android:layout_marginBottom="16dp"
+        android:layout_marginTop="16dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginStart="16dp" />
 </LinearLayout>
