--- app/src/main/res/layout/chamber_input.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal"
-    android:layout_marginTop="16dp"
-    android:layout_marginBottom="16dp"
-    android:gravity="center_vertical">
-
-    <SeekBar
-        android:id="@+id/seekBar"
-        android:layout_width="0dp"
-        android:layout_height="48dp"
-        android:layout_weight="1"
-        android:max="100"
-        android:minHeight="48dp" />
-
-    <EditText
-        android:id="@+id/editText"
-        android:layout_width="wrap_content"
-        android:layout_height="48dp"
-        android:minHeight="48dp"
-        android:layout_marginStart="16dp"
-        android:hint="Enter value"
-        android:inputType="numberSigned"
-        android:padding="8dp" />
-</LinearLayout>
