+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">FlexPAL GamePad</string>
     <string name="action_settings">Settings</string>
-    <string name="chamber_inputs">Chambers: %s</string>
+    <string name="chamber_inputs">Real-time inputs: %s</string>
     <string name="coordinates_text">Coordinates: %1$d, %2$d</string>
     <string name="title_text">FlexPAL GamePad</string>
     <!-- Strings used for fragments for navigation -->
